refactor: move sitekeys enum
