--- packages/apps/reputation-oracle/server/src/common/enums/site-key.ts
@@ -1,4 +1,0 @@
-export enum SiteKeyType {
-  HCAPTCHA = 'hcaptcha',
-  REGISTRATION = 'registration',
-}
