+++ packages/apps/reputation-oracle/server/src/modules/user/index.ts
@@ -2,5 +2,6 @@ export { UserEntity, UserStatus, Role as UserRole } from './user.entity';
 export { UserRepository } from './user.repository';
 export { UserService } from './user.service';
 export { UserModule } from './user.module';
+export { SiteKeyType } from './site-key.entity';
 
 export type * from './types';
